Move app/streams into app/domain/streams
And move classes under the Streams:: namespace
