+++ app/domain/streams/publishing_api/messages/single_item_message.rb
@@ -1,11 +1,11 @@
-module PublishingAPI
+module Streams::PublishingAPI
   class Messages::SingleItemMessage < Messages::BaseMessage
     def initialize(payload)
       super
     end
 
     def handler
-      PublishingAPI::Handlers::SingleItemHandler
+      Streams::PublishingAPI::Handlers::SingleItemHandler
     end
   end
 end
