+++ spec/domain/streams/publishing_api/consumer_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe PublishingAPI::Consumer do
+RSpec.describe Streams::PublishingAPI::Consumer do
   subject { described_class.new }
   it 'increments `discard` in statsd when routing key is not valid' do
     message = build(:message, routing_key: 'invalid_routing_key')
