+++ spec/domain/streams/publishing_api/message_processor_job_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe PublishingAPI::MessageProcessorJob do
+RSpec.describe Streams::PublishingAPI::MessageProcessorJob do
   it 'increments `monitor.messages.*` ' do
     message = build(:message, routing_key: 'news_story.major')
     expect(GovukStatsd).to receive(:increment).with("monitor.messages.major")
