+++ spec/domain/streams/publishing_api/messages/single_item_message_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe PublishingAPI::Messages::SingleItemMessage do
+RSpec.describe Streams::PublishingAPI::Messages::SingleItemMessage do
   subject { described_class }
   include_examples 'BaseMessage#historically_political?'
   include_examples 'BaseMessage#withdrawn_notice?'
