+++ spec/tasks/publishing_api_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe 'rake publishing_api:*', type: task do
   it 'starts listener for PublishingAPI events' do
     expect(GovukMessageQueueConsumer::Consumer).to receive(:new).with(
       queue_name: 'content_performance_manager',
-      processor: an_instance_of(PublishingAPI::Consumer),
+      processor: an_instance_of(Streams::PublishingAPI::Consumer),
     ).and_return(consumer)
 
     Rake::Task['publishing_api:consumer'].invoke
@@ -13,7 +13,7 @@ RSpec.describe 'rake publishing_api:*', type: task do
   it 'starts listener for PublishingAPI bulk preload events' do
     expect(GovukMessageQueueConsumer::Consumer).to receive(:new).with(
       queue_name: 'content_performance_manager_govuk_importer',
-      processor: an_instance_of(PublishingAPI::Consumer),
+      processor: an_instance_of(Streams::PublishingAPI::Consumer),
     ).and_return(consumer)
 
     Rake::Task['publishing_api:bulk_import_consumer'].invoke
